% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ytbh_nigh2015.R
\name{ytbh_nigh2015}
\alias{ytbh_nigh2015}
\title{Nigh (2015) years-to-breast-height model for Engelmann spruce}
\usage{
ytbh_nigh2015(si)
}
\arguments{
\item{si}{Numeric vector. Site index (m, base age 50 years at breast height).}
}
\value{
A tibble with one column:
\describe{
  \item{ytbh}{Predicted years to breast height (years).}
}
}
\description{
Implementation of the Nigh (2015) years-to-breast-height model for
Engelmann spruce (\code{PICE.ENG}) in the Engelmann Spruce -- Subalpine
Fir (ESSF) biogeoclimatic zone of British Columbia.
}
\details{
The source equation (model 2 in Nigh 2015) is:
\deqn{YTBH = 4.465 + 154.6 / SI}

This model was developed specifically for Engelmann spruce and is
recommended in place of the natural-stand white spruce years-to-breast-height
model that had previously been used for the species. It may be extrapolated
into zones neighbouring the ESSF provided the species is truly Engelmann
spruce; for the white x Engelmann cross, the white spruce models should be
used instead.
}
\examples{
ytbh_nigh2015(
  si = c(10, 15, 20)
)

}
\references{
Nigh, G. D. (2015). Years-to-breast-height model for Engelmann spruce in the
Engelmann Spruce -- Subalpine Fir biogeoclimatic zone. Province of British
Columbia, Victoria, B.C. Extension Note 115.
}
